requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
pytrends>=4.9.2
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
plotly>=5.18.0
python-dotenv>=1.0.0
